C search lago task 0

+++ 0x1E-search_algorithms/0-linear.c
@@ -17,16 +17,16 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t j;
+	int j;
 
 	if (array == NULL)
 	{
 		return (-1);
 	}
 
-	for (j = 0; j < size; j++)
+	for (j = 0; j < (int)size; j++)
 	{
-		printf("Value checked array[%li] = [%i]\n", j, array[j]);
+		printf("Value checked array[%u] = [%d]\n", j, array[j]);
 
 		if (array[j] == value)
 		{
